--- a/polars/src/chunked_array/ops/take.rs
+++ b/polars/src/chunked_array/ops/take.rs
@@ -7,10 +7,37 @@ use crate::chunked_array::builder::{
     get_list_builder, PrimitiveChunkedBuilder, Utf8ChunkedBuilder,
 };
 use crate::prelude::*;
+use crate::utils::Xob;
 use arrow::array::{
     Array, ArrayRef, BooleanArray, ListArray, PrimitiveArray, PrimitiveArrayOps, StringArray,
 };
+use arrow::compute::kernels::take::take;
 use std::sync::Arc;
+
+impl<T> ChunkedArray<T>
+where
+    T: PolarsDataType,
+{
+    pub fn take_from_single_chunked(&self, idx: &UInt32Chunked) -> Result<Self> {
+        if self.chunks.len() == 1 && idx.chunks.len() == 1 {
+            let idx_arr = idx.downcast_chunks()[0];
+
+            let a = &self.chunks[0];
+            let new_arr = take(a, idx_arr, None).unwrap();
+            Ok(Self::new_from_chunks(self.name(), vec![new_arr]))
+        } else {
+            Err(PolarsError::NoSlice)
+        }
+    }
+    pub fn take_from_single_chunked_iter(
+        &self,
+        indices: impl Iterator<Item = usize>,
+    ) -> Result<Self> {
+        let idx_ca: Xob<UInt32Chunked> = indices.into_iter().map(|idx| idx as u32).collect();
+        let idx_ca = idx_ca.into_inner();
+        self.take_from_single_chunked(&idx_ca)
+    }
+}
 
 macro_rules! impl_take_random_get {
     ($self:ident, $index:ident, $array_type:ty) => {{
@@ -181,6 +208,9 @@ where
     T: PolarsNumericType,
 {
     fn take(&self, indices: impl Iterator<Item = usize>, capacity: Option<usize>) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take!(self, indices, capacity, PrimitiveChunkedBuilder)
     }
 
@@ -189,6 +219,9 @@ where
         indices: impl Iterator<Item = usize>,
         capacity: Option<usize>,
     ) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take_unchecked!(self, indices, capacity, PrimitiveChunkedBuilder)
     }
 
@@ -214,6 +247,9 @@ impl ChunkTake for BooleanChunked {
     where
         Self: std::marker::Sized,
     {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take!(self, indices, capacity, PrimitiveChunkedBuilder)
     }
 
@@ -222,6 +258,9 @@ impl ChunkTake for BooleanChunked {
         indices: impl Iterator<Item = usize>,
         capacity: Option<usize>,
     ) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take_unchecked!(self, indices, capacity, PrimitiveChunkedBuilder)
     }
 
@@ -247,6 +286,9 @@ impl ChunkTake for Utf8Chunked {
     where
         Self: std::marker::Sized,
     {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take!(self, indices, capacity, Utf8ChunkedBuilder)
     }
 
@@ -255,6 +297,9 @@ impl ChunkTake for Utf8Chunked {
         indices: impl Iterator<Item = usize>,
         capacity: Option<usize>,
     ) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         impl_take_unchecked!(self, indices, capacity, Utf8ChunkedBuilder)
     }
 
@@ -280,6 +325,9 @@ impl ChunkTake for Utf8Chunked {
 
 impl ChunkTake for ListChunked {
     fn take(&self, indices: impl Iterator<Item = usize>, capacity: Option<usize>) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         let capacity = capacity.unwrap_or(indices.size_hint().0);
 
         match self.dtype() {
@@ -301,6 +349,9 @@ impl ChunkTake for ListChunked {
         indices: impl Iterator<Item = usize>,
         capacity: Option<usize>,
     ) -> Self {
+        if self.chunks.len() == 1 {
+            return self.take_from_single_chunked_iter(indices).unwrap();
+        }
         let capacity = capacity.unwrap_or(indices.size_hint().0);
         match self.dtype() {
             ArrowDataType::List(dt) => {
@@ -486,58 +537,22 @@ macro_rules! many_or_single {
     }};
 }
 
-pub enum NumTakeRandomDispatch<'a, T>
-where
-    T: PolarsNumericType,
-    T::Native: Copy,
-{
-    Cont(NumTakeRandomCont<'a, T::Native>),
-    Single(NumTakeRandomSingleChunk<'a, T>),
-    Many(NumTakeRandomChunked<'a, T>),
-}
-
-impl<'a, T> TakeRandom for NumTakeRandomDispatch<'a, T>
-where
-    T: PolarsNumericType,
-    T::Native: Copy,
-{
-    type Item = T::Native;
-
-    fn get(&self, index: usize) -> Option<Self::Item> {
-        use NumTakeRandomDispatch::*;
-        match self {
-            Cont(a) => a.get(index),
-            Single(a) => a.get(index),
-            Many(a) => a.get(index),
-        }
-    }
-
-    unsafe fn get_unchecked(&self, index: usize) -> Self::Item {
-        use NumTakeRandomDispatch::*;
-        match self {
-            Cont(a) => a.get_unchecked(index),
-            Single(a) => a.get_unchecked(index),
-            Many(a) => a.get_unchecked(index),
-        }
-    }
-}
-
 impl<'a, T> IntoTakeRandom<'a> for &'a ChunkedArray<T>
 where
     T: PolarsNumericType,
 {
     type Item = T::Native;
-    type TakeRandom = NumTakeRandomDispatch<'a, T>;
+    type TakeRandom = Box<dyn TakeRandom<Item = Self::Item> + 'a>;
 
     fn take_rand(&self) -> Self::TakeRandom {
         match self.cont_slice() {
-            Ok(slice) => NumTakeRandomDispatch::Cont(NumTakeRandomCont { slice }),
+            Ok(slice) => Box::new(NumTakeRandomCont { slice }),
             _ => {
                 let chunks = self.downcast_chunks();
                 if chunks.len() == 1 {
-                    NumTakeRandomDispatch::Single(NumTakeRandomSingleChunk { arr: chunks[0] })
+                    Box::new(NumTakeRandomSingleChunk { arr: chunks[0] })
                 } else {
-                    NumTakeRandomDispatch::Many(NumTakeRandomChunked { ca: self, chunks })
+                    Box::new(NumTakeRandomChunked { ca: self, chunks })
                 }
             }
         }