Noticed that taking DataFrame values by index used almost all operation time of joins when we have more than a few columns.

Some benchmark results

```
# Baseline
609 ms
600 ms
597 ms

## RAYON_NUM_THREADS=1
872 ms
902 ms
910 ms

# Only 1 column
25 ms
25 ms

# NO RAYON JOIN threads from join
590 ms
586 ms
589 ms

# Box<dyn TakeRand> instead of static dispatch per get
583 ms
575 ms
576 ms

# NO par_iter
852 ms
869 ms
873 ms

# Arrow take for UTF8
460 ms
461 ms
470 ms

# Arrow take for Boolean
589 ms
584 ms
578 ms
591 ms

# Arrow take for Primitive
581 ms
588 ms
594 ms


# Results of implements lowering the baseline
470 ms
477 ms
489 ms
```