@ritchie46 2020-11-29
> Interesting! What are the different benchmark results exactly?

I noticed that joins were much slower than expected when dealing with DataFrames with a lot of columns so I took a look at the indexing logic during a join.

The different things tried were:
* **RAYON_NUM_THREADS=1**
    - test if we parallelize too much.
* **NO par_iter**
    - test if we parallelize too much by changing the `par_iter` in the DataFrame indexing to `iter`.
* **Box<dyn TakeRand> instead of static dispatch per get**
  - [swap this enum] (https://github.com/ritchie46/polars/blob/133657cb00dd70fd1a967e6a0fce0cf6564fb414/polars/src/chunked_array/ops/take.rs#L489) with trait object.
* **Arrow take for UTF8 | Boolean | Primitive**
   - Use arrow's take kernel if ChunkedArray has got a single chunk.

Coincidently, you were also adding fast paths to the kernels so I vendored the snippet until next Arrow release. Only difference is that in the PrimitiveArray take I write directly to allocated memory.  I need to proper benchmark this, but I think that we should also do that in Arrow. Now the values are written to a `vec` and later copied to a buffer. That seems redundant. :thinking: