@Dandandan 2020-11-29
> but I think that we should also do that in Arrow. Now the values are written to a vec and later copied to a buffer

Yes, I think that's interesting. I think it is doing a memcpy from the slice eventually. Looks like that's happening with some more kernels, so probably enough low hanging fruit! I have to find out what exactly is the "current best way" of doing this in Arrow.