@ritchie46 2020-11-29
> Looks like that's happening with some more kernels, so probably enough low hanging fruit! I have to find out what exactly is the "current best way" of doing this in Arrow.

Indeed. I am also super curious about the discussion regarding `Vec<u8>` as an internal buffer/ dropping 64byte alignment requirement. No padding would make swapping from `Vec` or `ndarray` to arrow free. I will stay tuned. :)