@Dandandan 2020-11-30
https://github.com/apache/arrow/pull/8803 Looks like it has some substantial benefit indeed :)